Split Login into provider/controller
